refactor: Moved layer shifting service to common folder
