+++ projects/hslayers/src/common/layer-shifting/layer-shifting.service.mock.ts
@@ -1,0 +1,11 @@
+export function mockHsLayerShiftingService() {
+  return jasmine.createSpyObj('HsLayerShiftingService', [
+    'fillLayers',
+    'moveTo',
+    'getMaxZ',
+    'getMinZ',
+    'moveToBottom',
+    'moveToTop',
+    'swapSibling',
+  ]);
+}
