+++ projects/hslayers/src/common/layer-shifting/public-api.ts
@@ -1,0 +1,1 @@
+export * from './layer-shifting.service';
