+++ projects/hslayers/src/components/layermanager/physical-list/layermanager-physical-layerlist.component.ts
@@ -6,8 +6,8 @@ import {Subscription} from 'rxjs';
 import {HsConfig} from '../../../config.service';
 import {HsEventBusService} from '../../core/event-bus.service';
 import {HsLayerManagerService} from '../layermanager.service';
+import {HsLayerShiftingService} from '../../../common/layer-shifting/layer-shifting.service';
 import {HsLayerUtilsService} from '../../utils/layer-utils.service';
-import {HsLayermanagerPhysicalListService} from './layermanager-physical-layerlist.service';
 
 @Component({
   selector: 'hs-layermanager-physical-layer-list',
@@ -20,19 +20,18 @@ export class HsLayerPhysicalListComponent implements OnDestroy {
     public HsLayerManagerService: HsLayerManagerService,
     public HsLayerUtilsService: HsLayerUtilsService,
     public HsEventBusService: HsEventBusService,
-    public HsLayermanagerPhysicalListService: HsLayermanagerPhysicalListService,
+    public HsLayerShiftingService: HsLayerShiftingService,
     public HsConfig: HsConfig
   ) {
-    this.HsLayermanagerPhysicalListService.fillLayers();
+    this.HsLayerShiftingService.fillLayers();
     this.layerManagerUpdatesSubscription =
       this.HsEventBusService.layerManagerUpdates.subscribe((layer: any) => {
-        this.HsLayermanagerPhysicalListService.fillLayers();
+        this.HsLayerShiftingService.fillLayers();
         if (layer !== undefined) {
-          const layerFound =
-            this.HsLayermanagerPhysicalListService.layersCopy.find(
-              (wrapper) =>
-                wrapper.layer == layer || wrapper.layer == layer.layer
-            );
+          const layerFound = this.HsLayerShiftingService.layersCopy.find(
+            (wrapper) =>
+              wrapper.layer == layer || wrapper.layer == layer.layer
+          );
           if (layerFound !== undefined) {
             layerFound.active = true;
           }
@@ -47,14 +46,11 @@ export class HsLayerPhysicalListComponent implements OnDestroy {
     const replacedLayer = event.container.data[event.currentIndex];
 
     moveItemInArray(
-      this.HsLayermanagerPhysicalListService.layersCopy,
+      this.HsLayerShiftingService.layersCopy,
       event.previousIndex,
       event.currentIndex
     );
 
-    this.HsLayermanagerPhysicalListService.moveTo(
-      draggedLayer,
-      replacedLayer.layer
-    );
+    this.HsLayerShiftingService.moveTo(draggedLayer, replacedLayer.layer);
   }
 }
