+++ projects/hslayers/src/components/layermanager/physical-list/physical-layerlist.html
@@ -1,7 +1,7 @@
-<div class="list-group-item hs-lm-list pb-1" *ngIf="HsLayermanagerPhysicalListService.layersCopy.length != 0">
-    <div cdkDropList class="list-group row ps-1" (cdkDropListDropped)="drop($event)" [cdkDropListData]="HsLayermanagerPhysicalListService.layersCopy">
+<div class="list-group-item hs-lm-list pb-1" *ngIf="HsLayerShiftingService.layersCopy.length !== 0">
+    <div cdkDropList class="list-group row ps-1" (cdkDropListDropped)="drop($event)" [cdkDropListData]="HsLayerShiftingService.layersCopy">
         <div class="list-group-item hs-lm-item p-2" [ngClass]="{'activeLayer': layer.active}"
-            *ngFor="let layer of HsLayermanagerPhysicalListService.layersCopy;" cdkDrag>
+            *ngFor="let layer of HsLayerShiftingService.layersCopy;" cdkDrag>
                 <div class="d-flex align-items-center justify-content-between p-0 hs-lm-item-title flex-grow-1"
                     style="word-break:break-all; cursor:move">
                     <p class="m-0">{{HsLayerUtilsService.translateTitle(layer.title)}}</p>
